chore: move default migration handler to utils
